requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==2.0.3
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
lxml==3.8.0
numpy==1.13.1
packaging==16.8
pandas==0.20.3
pycparser==2.18
pyOpenSSL==17.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-ldap==2.4.25
pytz==2017.2
pyvmomi==6.5
PyYAML==3.11
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.12
xlrd==1.0.0
xlwt==1.2.0
